Fixed register to Login route and move creating user DB record to apiUserRoutes
