+++ controllers/login-routes.js
@@ -1,23 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const bcrypt = require('bcrypt');
-const mysql = require("mysql2");
 
 // const app = express();
 // app.use(express.json());
 
-const db = mysql.createConnection({
-    // put IP address of server instead of localhost
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PW,
-    database: process.env.DB_NAME
-});
-
-
-
-
-// /
 router.get('/',  (req, res) => {
     // console.log('slash is a wicked guitarist')
     res.render('home')
@@ -45,18 +31,8 @@ router.get('/register', (req, res) => {
 // }))
 // END HERE
 
-
-
 router.post('/register', async (req, res) => {
     try {
-        const hashedPassword = await bcrypt.hash(req.body.password, 10)
-        users.push({
-            id: Date.now().toString(),
-            //id for the database
-            name: req.body.name,
-            email: req.body.email,
-            password: hashedPassword
-        })
         res.redirect('/login')
     } catch {
         res.redirect('/register')
@@ -64,7 +40,10 @@ router.post('/register', async (req, res) => {
     // console.log(users)
 })
 
-// db.add(users);
+
+
+
+
 
 router.delete('/logout', (req, res) => {
     req.logOut()
